Refactoring type hash to other file
